backend/requirements.txt
# --- API framework & ASGI ---
fastapi==0.121.0
fastapi-cli==0.0.16
uvicorn[standard]==0.38.0
python-multipart==0.0.9  # 파일 업로드(multipart/form-data) 지원
httpx==0.27.0  # 비동기 HTTP 클라이언트 (LLM API 호출용)

# --- Database ---
sqlalchemy==2.0.44
sqlalchemy2-stubs==0.0.2a38
psycopg2-binary==2.9.9 # PostgreSQL 드라이버
pgvector==0.2.3

# --- Embedding & ML ---
sentence-transformers>=2.7.0
torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu
numpy>=1.24.0

# --- Configuration & validation ---
python-dotenv==1.2.1
pydantic==2.12.3
pydantic-settings==2.6.0
email-validator==2.1.1  # Pydantic EmailStr 타입을 위한 이메일 검증 라이브러리

# --- Security ---
PyJWT==2.9.0

# --- AI & imaging ---
google-genai==0.3.0
Pillow==10.4.0
requests==2.31.0  # 이미지 다운로드용
mediapipe==0.10.14  # 포즈 검증용

# --- Async file I/O ---
aiofiles==24.1.0

# --- Scheduling ---
APScheduler==3.10.4

boto3>=1.28.0